+++ ZcashLightClientKitTests/OutboundTransactionManagerTests.swift
@@ -19,7 +19,7 @@ class OutboundTransactionManagerTests: XCTestCase {
     var initializer: Initializer!
     let spendingKey = "secret-extended-key-test1qvpevftsqqqqpqy52ut2vv24a2qh7nsukew7qg9pq6djfwyc3xt5vaxuenshp2hhspp9qmqvdh0gs2ljpwxders5jkwgyhgln0drjqaguaenfhehz4esdl4kwlm5t9q0l6wmzcrvcf5ed6dqzvct3e2ge7f6qdvzhp02m7sp5a0qjssrwpdh7u6tq89hl3wchuq8ljq8r8rwd6xdwh3nry9at80z7amnj3s6ah4jevnvfr08gxpws523z95g6dmn4wm6l3658kd4xcq9rc0qn"
     let recipientAddress = "ztestsapling1ctuamfer5xjnnrdr3xdazenljx0mu0gutcf9u9e74tr2d3jwjnt0qllzxaplu54hgc2tyjdc2p6"
-    let zpend: Int = 500_000
+    let zpend: Int64 = 500_000
     
     override func setUp() {
         
@@ -63,7 +63,7 @@ class OutboundTransactionManagerTests: XCTestCase {
         
         XCTAssertEqual(pendingTx.toAddress, recipientAddress)
         XCTAssertEqual(pendingTx.memo, nil)
-        XCTAssertEqual(pendingTx.value, zpend)
+        XCTAssertEqual(pendingTx.value, Int(zpend))
         
     }
